Trying to deal with negative trial stresses returned by laromance mode
Look for fixme lynn comments on what was done to laromance model with
Ben and Stephanie.

- tried to set _check_range = true   so radialReturn would use bisection.
This gets rid of the negative trial stress but the radial return does
not converge after 1000 iterations.

- Trying to modify the rom_effective_strain but this too is not fixing the
issues we are having. We may need to also modify the other quantities being
updated in the larom model when it is outside the window bounds of the model.
Maybe we need a variable to smoothly vary the larom model to either return full
elastic or full creep depending on which side of the window it is exceeding

- Added a 2D axis symmetric pipe model that I am using for the NEML stochastic simulations
There are elastic versions of this and then when I add LAROM to the simulation it works for
a while but then crashes.  We would want to run this simulation for 20 years.  I've been able
to run it for about 80 days.  The version that is in there won't run very long before it fails
to converge from going outside the bounds. This might be due using a reference residual or
setting the nonlinear tolerance too tight.  I also have the neml model input file in here with
its xml just for reference.  This only runs with blackbear so it should be deleted before
pushing.