change rom error to exception in hopes of getting past bad values with a smaller timestep. closes #17080
This will drive the timestep down to dtmin and then the exception will trigger an error but in a sampler,
driving the the timestep down to dtmin causes the multapp to abort which lets the multiapp continue collecting
samples if ignore_solve_not_converge = true is set in the SamplerFullSolveMultiApp.  If we hit mooseErrors,
the multiApp would be terminated and we would lose all of the samples.